Added

+++ apps/satin/fib/FibInterface.java
@@ -1,3 +1,3 @@
 interface FibInterface extends ibis.satin.Spawnable {
-	public long fib(long n);
+	public long fib(int n);
 }
